+++ DalaMock/Mocks/MockReplacementContainer.cs
@@ -3,12 +3,17 @@ namespace DalaMock.Core.Mocks;
 /// <inheritdoc />
 public class MockReplacementContainer : IReplacementContainer
 {
-    public MockReplacementContainer(IUiBuilder uiBuilder, MockWindowSystem.Factory factory)
+    public MockReplacementContainer(
+        IUiBuilder uiBuilder,
+        MockWindowSystem.Factory factory,
+        MockDalamudConfiguration dalamudConfiguration,
+        MockSystemFontProvider systemFontProvider)
     {
         this.ImGuiComponents = new MockImGuiComponents(uiBuilder);
         this.WindowSystemFactory = new MockWindowSystemFactory(factory);
         this.Font = new MockFont();
         this.FileDialogManager = new MockFileDialogManager();
+        this.FontChooserFactory = new MockFontChooserFactory(uiBuilder, dalamudConfiguration, systemFontProvider);
     }
 
     public IImGuiComponents ImGuiComponents { get; }
@@ -19,6 +24,8 @@ public class MockReplacementContainer : IReplacementContainer
 
     public IFileDialogManager FileDialogManager { get; }
 
+    public IFontChooserFactory FontChooserFactory { get; }
+
     /// <inheritdoc/>
     public void Register(ContainerBuilder containerBuilder)
     {
@@ -26,6 +33,7 @@ public class MockReplacementContainer : IReplacementContainer
         containerBuilder.RegisterInstance(this.WindowSystemFactory).AsImplementedInterfaces().AsSelf().SingleInstance();
         containerBuilder.RegisterInstance(this.Font).AsImplementedInterfaces().AsSelf().SingleInstance();
         containerBuilder.RegisterInstance(this.FileDialogManager).AsImplementedInterfaces().AsSelf().SingleInstance();
+        containerBuilder.RegisterInstance(this.FontChooserFactory).AsImplementedInterfaces().AsSelf().SingleInstance();
         containerBuilder.RegisterType<MockWindowSystem>().As<IWindowSystem>().InstancePerDependency();
     }
 }