+++ CHANGELOG.md
@@ -13,6 +13,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 
+## [6.1.0] - 2026-05-24
+
+### Added
+- Added custom font/style support to DalaMock.
+- Added full support for game fonts + ttf fonts
+- Added IFontChooserFactory and made it available in IReplacementContainer
+- Added a font test window to the sample.
+- Added ability to set the global ui scale, style and default font which are currently saved locally per dalamock instance.
+
 ## [6.0.1] - 2026-05-10
 
 ### Added